spl_token_interface::state

Struct Account

#[repr(C)]
pub struct Account {
    pub mint: Pubkey,
    pub owner: Pubkey,
    pub amount: u64,
    pub delegate: COption<Pubkey>,
    pub state: AccountState,
    pub is_native: COption<u64>,
    pub delegated_amount: u64,
    pub close_authority: COption<Pubkey>,
}

Account data.

Fields
mint: Pubkey

The mint associated with this account

owner: Pubkey

The owner of this account.

amount: u64

The amount of tokens this account holds.

delegate: COption<Pubkey>

If delegate is Some then delegated_amount represents the amount authorized by the delegate

state: AccountState

The account’s state

is_native: COption<u64>

If is_native.is_some, this is a native token, and the value logs the rent-exempt reserve. An Account is required to be rent-exempt, so the value is used by the Processor to ensure that wrapped SOL accounts do not drop below this threshold.

delegated_amount: u64

The amount delegated

close_authority: COption<Pubkey>

Optional authority to close the account.

Implementations

impl Account

pub fn is_frozen(&self) -> bool

Checks if account is frozen

pub fn is_native(&self) -> bool

Checks if account is native

pub fn is_owned_by_system_program_or_incinerator(&self) -> bool

Checks if a token Account’s owner is the system_program or the incinerator